feat: paralellize download

+++ src/consts.rs
@@ -1,4 +1,4 @@
 pub const VALIDATOR_AS_BUFFER: &[u8] = "Validator".as_bytes();
 pub const BUNDLR_AS_BUFFER: &[u8] = "Bundlr".as_bytes();
 
-pub const CHUNK_SIZE : usize = 256 * 1024;      //256kb
+pub const CHUNK_SIZE : u64 = 256 * 1024;      //256kb
